Move arcs-make and makeTSVfile.py to bin
